requirements.txt
midiutils
pygame>=2.0.0.dev6
more-itertools